+++ brent-measures_uk.csv
@@ -1,1 +1,21 @@
-#time,date,susceptible,exposed,infectious,recovered,dead,immune,num infections today,num hospitalisations today,hospital bed occupancy,num hospitalisations today (data)
+#time,date,susceptible,exposed,infectious,recovered,dead,immune,num infections today,num hospitalisations today,hospital bed occupancy,num hospitalisations today (data),cum num hospitalisations today,cum num infections today
+1,2/3/2020,324542,2985,1732,870,0,0,0,0,17,0.0,0,0
+2,3/3/2020,323443,3471,2136,1079,0,0,0,0,20,0.0,0,0
+3,4/3/2020,322022,4131,2636,1340,0,0,0,0,23,0.0,0,0
+4,5/3/2020,320269,4995,3209,1656,0,0,0,0,28,0.0,0,0
+5,6/3/2020,318253,5977,3850,2047,2,0,0,0,36,0.0,0,0
+6,7/3/2020,315929,7051,4573,2573,3,0,0,0,50,0.0,0,0
+7,8/3/2020,313306,8155,5460,3205,3,0,0,0,59,0.0,0,0
+8,9/3/2020,310317,9446,6433,3930,3,0,0,0,78,0.0,0,0
+9,10/3/2020,307002,10722,7643,4758,4,0,0,0,89,0.0,0,0
+10,11/3/2020,303075,12310,8987,5753,4,0,0,0,115,0.0,0,0
+11,12/3/2020,298825,13906,10424,6968,6,0,0,0,134,0.0,0,0
+12,13/3/2020,295048,14574,12105,8396,6,0,0,0,158,0.0,0,0
+13,14/3/2020,290928,15340,13802,10052,7,0,0,0,194,0.0,0,0
+14,15/3/2020,286345,16245,15572,11959,8,0,0,0,250,0.0,0,0
+15,16/3/2020,281508,17257,17186,14168,10,0,0,0,291,0.0,0,0
+16,17/3/2020,278340,16376,18699,16701,13,0,0,0,347,0.0,0,0
+17,18/3/2020,274956,15543,20168,19448,14,0,0,0,429,0.0,0,0
+18,19/3/2020,271476,15030,21015,22590,18,0,0,0,511,0.0,0,0
+19,20/3/2020,267890,14683,21584,25953,19,0,0,0,608,0.0,0,0
+20,21/3/2020,265831,12942,21885,29445,26,0,0,0,699,0.0,0,0
