+++ facs/base/facs.py
@@ -16,7 +16,8 @@ import pandas as pd
 
 from .needs import Needs
 from .location_types import building_types_dict, building_types
-from .utils import probability, get_random_int
+from .person import Person
+from .utils import probability, get_random_int, out_files
 
 try:
     from mpi4py import MPI
@@ -82,26 +83,6 @@ num_deaths_today = 0
 num_recoveries_today = 0
 
 
-class OUTPUT_FILES:
-    def __init__(self):
-        self.files = {}
-
-    def open(self, file_name):
-        if not file_name in self.files:
-            if os.path.exists(file_name):
-                os.remove(file_name)
-            self.files[file_name] = open(file_name, "a")
-
-        return self.files[file_name]
-
-    def __del__(self) -> None:
-        for out_file in self.files:
-            self.files[out_file].close()
-
-
-out_files = OUTPUT_FILES()
-
-
 def write_log_headers(rank):
     out_inf = out_files.open("{}/covid_out_infections_{}.csv".format(log_prefix, rank))
     print("#time,x,y,location_type,rank,incubation_time", file=out_inf, flush=True)
@@ -113,281 +94,6 @@ def write_log_headers(rank):
     print("#time,x,y,age", file=out_inf, flush=True)
     out_inf = out_files.open("{}/covid_out_recoveries_{}.csv".format(log_prefix, rank))
     print("#time,x,y,age", file=out_inf, flush=True)
-
-
-def log_infection(t, x, y, loc_type, rank, phase_duration):
-    global num_infections_today
-    out_inf = out_files.open("{}/covid_out_infections_{}.csv".format(log_prefix, rank))
-    print(
-        "{},{},{},{},{},{}".format(t, x, y, loc_type, rank, phase_duration),
-        file=out_inf,
-        flush=True,
-    )
-    num_infections_today += 1
-
-
-def log_hospitalisation(t, x, y, age, rank):
-    global num_hospitalisations_today
-    out_inf = out_files.open(
-        "{}/covid_out_hospitalisations_{}.csv".format(log_prefix, rank)
-    )
-    print("{},{},{},{}".format(t, x, y, age), file=out_inf, flush=True)
-    num_hospitalisations_today += 1
-
-
-def log_death(t, x, y, age, rank):
-    global num_deaths_today
-    out_inf = out_files.open("{}/covid_out_deaths_{}.csv".format(log_prefix, rank))
-    print("{},{},{},{}".format(t, x, y, age), file=out_inf, flush=True)
-    num_deaths_today += 1
-
-
-def log_recovery(t, x, y, age, rank):
-    global num_recoveries_today
-    out_inf = out_files.open("{}/covid_out_recoveries_{}.csv".format(log_prefix, rank))
-
-    print("{},{},{},{}".format(t, x, y, age), file=out_inf, flush=True)
-    num_recoveries_today += 1
-
-
-class Person:
-    def __init__(self, location, household, ages):
-        self.location = location  # current location
-        self.location.IncrementNumAgents()
-        self.home_location = location
-        self.household = household
-        self.mild_version = True
-        self.hospitalised = False
-        self.dying = False
-        self.work_from_home = False
-        self.school_from_home = False
-        self.phase_duration = 0.0  # duration to next phase.
-        self.symptoms_suppressed = (
-            False  # Symptoms suppressed, e.g. due to vaccination, but still infectious.
-        )
-        self.antivax = False  # Refuses vaccines if true.
-        if np.random.rand() < 0.05:  # 5% are antivaxxers.
-            self.antivax = True
-
-        self.status = "susceptible"  # states: susceptible, exposed, infectious, recovered, dead, immune.
-        self.symptomatic = False  # may be symptomatic if infectious
-        self.status_change_time = -1
-
-        self.age = np.random.choice(91, p=ages)  # age in years
-        self.job = np.random.choice(4, 1, p=[0.865, 0.015, 0.08, 0.04])[
-            0
-        ]  # 0=default, 1=teacher (1.5%), 2=shop worker (8%), 3=health worker (4%)
-
-    def assign_group(self, location_type, num_groups):
-        """
-        Used to assign a grouping to a person.
-        For example, a campus may have 30 classes (num_groups = 30). Then you would use:
-        assign_group("school", 30)
-        The location type should match the corresponding personal needs category (e.g., school or supermarket).
-        """
-        if not hasattr(self, "groups"):
-            self.groups = {}
-        self.groups[building_types_dict[location_type]] = get_random_int(num_groups)
-
-    def location_has_grouping(self, lid):
-        """
-        Takes a location type ID: return True or False
-        """
-        if hasattr(self, "groups"):
-            if lid in list(self.groups.keys()):
-                # print(lid, list(self.groups.keys()))
-                return True
-        return False
-
-    def vaccinate(self, time, vac_no_symptoms, vac_no_transmission, vac_duration):
-        self.status_change_time = time  # necessary if vaccines give temporary immunity.
-        if vac_duration > 0:
-            if vac_duration > 100:
-                self.phase_duration = np.random.gamma(
-                    vac_duration / 20.0, 20.0
-                )  # shape parameter is changed with variable, scale parameter is kept fixed at 20 (assumption).
-
-            else:
-                self.phase_duration = np.poisson(vac_duration)
-        if self.status == "susceptible":
-            if probability(vac_no_transmission):
-                self.status = "immune"
-            elif probability(vac_no_symptoms):
-                self.symptoms_suppressed = True
-        # print("vac", self.status, self.symptoms_suppressed, self.phase_duration)
-
-    def plan_visits(self, e, deterministic=False):
-        """
-        Plan visits for the day.
-        TODO: plan visits to classes not using nearest location (make an override).
-        """
-
-        if self.status in [
-            "susceptible",
-            "exposed",
-            "infectious",
-        ]:  # recovered people are assumed to be immune.
-            personal_needs = needs.get_needs(self)
-            for k, minutes in enumerate(personal_needs):
-                nearest_locs = self.home_location.nearest_locations
-
-                if minutes < 1:
-                    continue
-                elif k == building_types_dict["hospital"] and self.hospitalised:
-                    location_to_visit = self.hospital
-
-                elif k == building_types_dict["office"] and self.job > 0:
-                    if self.job == 1:  # teacher
-                        location_to_visit = nearest_locs[building_types_dict["school"]]
-                    if self.job == 2:  # shop employee
-                        location_to_visit = nearest_locs[
-                            building_types_dict["shopping"]
-                        ]
-                    if self.job == 3:  # health worker
-                        location_to_visit = nearest_locs[
-                            building_types_dict["hospital"]
-                        ]
-
-                elif self.location_has_grouping(k):
-                    location_to_visit = e.get_location_by_group(k, self.groups[k])
-
-                elif nearest_locs[k]:
-                    location_to_visit = nearest_locs[k]
-                else:  # no known nearby locations.
-                    continue
-
-                e.visit_minutes += minutes
-                location_to_visit.register_visit(e, self, minutes, deterministic)
-
-    def print_needs(self):
-        print(self.age, needs.get_needs(self))
-
-    def get_needs(self):
-        return needs.get_needs(self)
-
-    def get_hospitalisation_chance(self, disease):
-        age = int(min(self.age, len(disease.hospital) - 1))
-        return disease.hospital[age]
-
-    def get_mortality_chance(self, disease):
-        age = int(min(self.age, len(disease.hospital) - 1))
-        return disease.mortality[age]
-
-    def infect(self, e, severity="exposed", location_type="house"):
-        # severity can be overridden to infectious when rigidly inserting cases.
-        # but by default, it should be exposed.
-        self.status = severity
-        self.status_change_time = e.time
-        self.mild_version = True
-        self.hospitalised = False
-        self.phase_duration = max(1, np.random.poisson(e.disease.incubation_period))
-        log_infection(
-            e.time,
-            self.location.x,
-            self.location.y,
-            location_type,
-            e.rank,
-            self.phase_duration,
-        )
-
-    def recover(self, e, location):
-        if e.disease.immunity_duration > 0:
-            self.phase_duration = np.random.gamma(
-                e.disease.immunity_duration / 20.0, 20.0
-            )  # shape parameter is changed with variable, scale parameter is kept fixed at 20 (assumption).
-        self.status = "recovered"
-        self.status_change_time = e.time
-        log_recovery(e.time, self.location.x, self.location.y, location, e.rank)
-
-    def progress_condition(self, e, t, disease):
-        if self.status_change_time > t:
-            return
-        if self.status == "exposed":
-            # print("exposed", t, self.status_change_time, self.phase_duration)
-            if t - self.status_change_time >= int(self.phase_duration):
-                self.status = "infectious"
-                self.status_change_time = t
-                if (
-                    probability(self.get_hospitalisation_chance(disease))
-                    and self.symptoms_suppressed == False
-                ):
-                    self.mild_version = False
-                    # self.phase_duration = np.random.poisson(disease.period_to_hospitalisation - disease.incubation_period)
-                    self.phase_duration = max(
-                        1,
-                        np.random.poisson(disease.period_to_hospitalisation)
-                        - self.phase_duration,
-                    )
-                else:
-                    self.mild_version = True
-                    # self.phase_duration = np.random.poisson(disease.mild_recovery_period - disease.incubation_period)
-                    self.phase_duration = max(
-                        1,
-                        np.random.poisson(disease.mild_recovery_period)
-                        - self.phase_duration,
-                    )
-
-        elif self.status == "infectious":
-            # mild version (may require hospital visits, but not ICU visits)
-            if self.mild_version:
-                if t - self.status_change_time >= self.phase_duration:
-                    self.recover(e, "house")
-
-            # non-mild version (will involve ICU visit)
-            else:
-                if not self.hospitalised:
-                    if t - self.status_change_time >= self.phase_duration:
-                        self.hospitalised = True
-                        self.hospital = e.find_hospital()
-                        if self.hospital == None:
-                            print(
-                                "Error: agent is hospitalised, but there are no hospitals in the location graph."
-                            )
-                            sys.exit()
-                        e.num_hospitalised += 1
-                        log_hospitalisation(
-                            t, self.location.x, self.location.y, self.age, e.rank
-                        )
-
-                        self.status_change_time = t  # hospitalisation is a status change, because recovery_period is from date of hospitalisation.
-                        if probability(
-                            self.get_mortality_chance(disease)
-                            / self.get_hospitalisation_chance(disease)
-                        ):
-                            # avg mortality rate (divided by the average hospitalization rate). TODO: read from YML.
-                            self.dying = True
-                            self.phase_duration = np.random.poisson(
-                                disease.mortality_period
-                            )
-                        else:
-                            self.dying = False
-                            self.phase_duration = np.random.poisson(
-                                disease.recovery_period
-                            )
-                else:
-                    if (
-                        t - self.status_change_time >= self.phase_duration
-                    ):  # from hosp. date
-                        self.hospitalised = False
-                        e.num_hospitalised -= 1
-                        self.status_change_time = t
-                        # decease
-                        if self.dying:
-                            self.status = "dead"
-                            log_death(
-                                t, self.location.x, self.location.y, "hospital", e.rank
-                            )
-                        # hospital discharge
-                        else:
-                            self.recover(e, "hospital")
-
-        elif e.disease.immunity_duration > 0 and (
-            self.status == "recovered" or self.status == "immune"
-        ):
-            if t - self.status_change_time >= self.phase_duration:
-                # print("susc.", self.status, self.phase_duration)
-                self.status = "susceptible"
-                self.symptoms_suppressed = False
 
 
 class Household:
